+++ packages/design-system/src/components/Stack/StackHorizontal.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import tokens from '@talend/design-tokens';
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { StackHorizontal } from './index';
+import { StackHorizontal } from '.';
 import {
 	alignContentOptions,
 	alignOptions,
@@ -16,6 +16,7 @@ export default {
 	component: StackHorizontal,
 } as ComponentMeta<typeof StackHorizontal>;
 
+// eslint-disable-next-line storybook/prefer-pascal-case
 export const manualStackArgs = {
 	as: { options: [...possibleAsTypes], control: { type: 'select' }, defaultValue: 'div' },
 	justify: {
@@ -59,7 +60,7 @@ function Block({ width }: { width: string }) {
 				height: tokens.coralSizeL,
 				borderRadius: tokens.coralRadiusM,
 				background: tokens.coralColorAccentBackground,
-				border: `${tokens.coralBorderDashedS} ${tokens.coralColorAccentBorder}`,
+				border: `${tokens.coralBorderSDashed} ${tokens.coralColorAccentBorder}`,
 			}}
 		/>
 	);